# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Configurations API",
        "description": "API for managing cars, configurations, options and prices",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/configurations": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Get all configurations",
                "operationId": "a28c5d89c44060df704ee25a16aab614",
                "responses": {
                    "200": {
                        "description": "List of all configurations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Create a new configuration",
                "operationId": "9a8560ffccc4628bcf83ade91fb34040",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "name"
                                ],
                                "properties": {
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Comfort"
                                    },
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Configuration created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/configurations/{id}": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Get a single configuration",
                "operationId": "751d7e481c28c75b8e84b9b9d1369a52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the configuration to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration details",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Update a configuration",
                "operationId": "c813010747e3179533ea014b3f892a4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the configuration to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "name"
                                ],
                                "properties": {
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Comfort"
                                    },
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Configuration not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Delete a configuration",
                "operationId": "e7ce276f1c55cb033607295b38381a5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the configuration to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Configuration deleted successfully"
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                }
            }
        },
...
    },
...
}